+++ Editor/CityImport/PackageLoadConfigGUIs/GUIParts/FallbackMaterialGUI.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: f3aedc83209e4420ae0a730f16fff9bb
+timeCreated: 1698639144
