+++ Editor/CityImport/PackageLoadConfigGUIs/GUIParts/PositionConfGUI.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: f6a357b0ebe5415795cdd95dfafba219
+timeCreated: 1698639238
