+++ Editor/CityImport/PackageLoadConfigGUIs/PackageLoadConfigGUI.cs
@@ -1,13 +1,12 @@
-using System;
 using System.Collections.Generic;
 using PLATEAU.CityImport.Config;
+using PLATEAU.CityImport.Config.PackageLoadConfigs;
 using PLATEAU.Dataset;
-using PLATEAU.Editor.CityImport.GUIParts;
+using PLATEAU.Editor.CityImport.PackageLoadConfigGUIs.GUIParts;
 using PLATEAU.Editor.EditorWindow.Common;
 using UnityEditor;
-using UnityEngine;
 
-namespace PLATEAU.Editor.CityImport.PackageLodSettingGUIs
+namespace PLATEAU.Editor.CityImport.PackageLoadConfigGUIs
 {
     /// <summary>
     /// インポート設定のうち、パッケージ種1つの設定GUIです。
@@ -67,62 +66,5 @@ namespace PLATEAU.Editor.CityImport.PackageLodSettingGUIs
             // サブクラスで実装しない限り何もしません
         }
     
-    }
-
-    /// <summary>
-    /// パッケージごとの設定GUIについて、各設定項目ごとのGUIをコンポーネントと呼ぶことにします。
-    /// 例えば建物のテクスチャに関する設定コンポーネント、道路のメッシュ結合粒度に関する設定コンポーネント、といったものです。
-    /// これら各コンポーネントは、 PackageLoadSetting を保持し、 Draw() で設定GUIを描画し、
-    /// その中でユーザーのGUI操作に応じて設定値を変更します。
-    /// そのコンポーネントという概念を抽象化したものがこのクラスです。
-    /// </summary>
-    internal abstract class PackageLoadConfigGUIComponent : IEditorDrawable
-    {
-        protected readonly PackageLoadConfig Conf;
-
-        protected PackageLoadConfigGUIComponent(PackageLoadConfig conf)
-        {
-            Conf = conf;
-        }
-        public abstract void Draw();
-        public void Dispose() { }
-    }
-
-    
-    /// <summary>
-    /// LOD範囲の設定GUIです。
-    /// </summary>
-    internal class LodGUI : PackageLoadConfigGUIComponent
-    {
-        public LodGUI(PackageLoadConfig conf) : base(conf)
-        {
-        }
-
-        public override void Draw()
-        {
-            var predefined = CityModelPackageInfo.GetPredefined(Conf.Package);
-            int minLOD = Conf.LODRange.MinLOD;
-            int maxLOD = Conf.LODRange.MaxLOD;
-            int availableMaxLOD = Conf.LODRange.AvailableMaxLOD;
-            PlateauEditorStyle.LODSlider("LOD描画設定", ref minLOD, ref maxLOD,
-                Math.Min(predefined.minLOD, availableMaxLOD), availableMaxLOD);
-            Conf.LODRange = new LODRange(minLOD, maxLOD, availableMaxLOD);
-        }
-    }
-
-    /// <summary>
-    /// デフォルトマテリアルの設定GUIです。
-    /// </summary>
-    internal class FallbackMaterialGUI : PackageLoadConfigGUIComponent
-    {
-        public FallbackMaterialGUI(PackageLoadConfig conf) : base(conf)
-        {
-        }
-
-        public override void Draw()
-        {
-            Conf.FallbackMaterial = (Material)EditorGUILayout.ObjectField("デフォルトマテリアル",
-                Conf.FallbackMaterial, typeof(Material), false);
-        }
     }
 }
