+++ Editor/CityImport/PackageLoadConfigGUIs/ReliefLoadConfigGUI.cs
@@ -1,14 +1,15 @@
 using System;
 using PLATEAU.CityImport.Config;
+using PLATEAU.CityImport.Config.PackageLoadConfigs;
 using PLATEAU.Dataset;
-using PLATEAU.Editor.CityImport.GUIParts;
+using PLATEAU.Editor.CityImport.PackageLoadConfigGUIs.GUIParts;
 using PLATEAU.Editor.EditorWindow.Common;
 using UnityEditor;
 using UnityEngine;
 #if UNITY_EDITOR
 #endif
 
-namespace PLATEAU.Editor.CityImport.PackageLodSettingGUIs
+namespace PLATEAU.Editor.CityImport.PackageLoadConfigGUIs
 {
     /// <summary>
     /// <see cref="ReliefLoadConfig"/> に対応するGUIクラスです。
