+++ Editor/EditorWindow/PlateauWindow/MainTabGUI/ImportGUIParts/ImportGUIAfterAreaSelect.cs
@@ -1,4 +1,5 @@
 ﻿using PLATEAU.CityImport.Config;
+using PLATEAU.CityImport.Config.PackageLoadConfigs;
 using PLATEAU.Editor.CityImport;
 using PLATEAU.Editor.EditorWindow.Common;
 using PLATEAU.Editor.EditorWindow.ProgressDisplay;
