+++ Runtime/CityImport/AreaSelector/AreaSelectorDataPass.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using PLATEAU.CityImport.Config;
+using PLATEAU.CityImport.Config.PackageLoadConfigs;
 using PLATEAU.Dataset;
 using UnityEngine;
 using UnityEngine.SceneManagement;
