+++ Runtime/CityImport/AreaSelector/IAreaSelectResultReceiver.cs
@@ -1,4 +1,5 @@
 ﻿using PLATEAU.CityImport.Config;
+using PLATEAU.CityImport.Config.PackageLoadConfigs;
 
 namespace PLATEAU.CityImport.AreaSelector
 {
