+++ Runtime/CityImport/Config/PackageLoadConfigs/PackageLoadConfig.cs
@@ -1,11 +1,10 @@
-﻿using System;
-using PLATEAU.Dataset;
+﻿using PLATEAU.Dataset;
 using PLATEAU.Geometries;
 using PLATEAU.Native;
 using PLATEAU.PolygonMesh;
 using UnityEngine;
 
-namespace PLATEAU.CityImport.Config
+namespace PLATEAU.CityImport.Config.PackageLoadConfigs
 {
 
     /// <summary>
@@ -159,94 +158,6 @@ namespace PLATEAU.CityImport.Config
         private static bool ShouldExcludePolygonsOutsideExtent(PredefinedCityModelPackage package)
         {
             return !ShouldExcludeCityObjectOutsideExtent(package);
-        }
-    }
-
-    /// <summary>
-    /// パッケージごとの設定項目に、土地特有の設定項目を追加したクラスです。
-    /// これに対応するGUIクラスは ReliefLoadSettingGUI を参照してください。
-    /// </summary>
-    internal class ReliefLoadConfig : PackageLoadConfig
-    {
-        public bool AttachMapTile { get; set; }
-
-        private int mapTileZoomLevel;
-        public const int MinZoomLevel = 1;
-
-        /// <summary> 地理院地図やGoogleMapPlatformで存在するズームレベルの最大値です。2023年8月に調べたところ18でした。 </summary>
-        public const int MaxZoomLevel = 18;
-
-        public const string DefaultMapTileUrl = "https://cyberjapandata.gsi.go.jp/xyz/seamlessphoto/{z}/{x}/{y}.jpg";
-
-        public int MapTileZoomLevel
-        {
-            get => this.mapTileZoomLevel;
-            set
-            {
-                if (value < MinZoomLevel || value > MaxZoomLevel)
-                {
-                    throw new ArgumentOutOfRangeException(nameof(MapTileZoomLevel));
-                }
-
-                this.mapTileZoomLevel = value;
-            }
-        }
-
-        private string mapTileURL;
-
-        public string MapTileURL
-        {
-            get => this.mapTileURL;
-            set
-            {
-                if ((!value.StartsWith("http://")) && (!value.StartsWith("https://")))
-                    throw new ArgumentException("URL must start with https:// or http://.");
-                this.mapTileURL = value;
-            }
-        }
-
-        public ReliefLoadConfig(PackageLoadConfig baseConfig) :
-            base(baseConfig)
-        {
-            // 初期値を決めます。
-            AttachMapTile = true;
-            MapTileZoomLevel = 17;
-            MapTileURL = DefaultMapTileUrl;
-        }
-
-        public override MeshExtractOptions ConvertToNativeOption(PlateauVector3d referencePoint, int coordinateZoneID)
-        {
-            var nativeOption = base.ConvertToNativeOption(referencePoint, coordinateZoneID);
-            nativeOption.AttachMapTile = AttachMapTile;
-            nativeOption.MapTileZoomLevel = MapTileZoomLevel;
-            nativeOption.MapTileURL = MapTileURL;
-            nativeOption.TexturePackingResolution = GetTexturePackingResolution();
-            return nativeOption;
-        }
-
-        private uint GetTexturePackingResolution()
-        {
-            return ConfExtendable.GetTexturePackingResolution();
-        }
-    }
-
-    internal struct LODRange
-    {
-        /// <summary> ユーザーが選択したLOD範囲の下限 </summary>
-        public int MinLOD { get; }
-
-        /// <summary> ユーザーが選択したLODの上限 </summary>
-        public int MaxLOD { get; }
-
-        /// <summary> ユーザーが選択したデータのなかで存在するLODの最大値 </summary>
-        public int AvailableMaxLOD { get; }
-
-        public LODRange(int minLOD, int maxLOD, int availableMaxLOD)
-        {
-            if (minLOD > maxLOD) throw new ArgumentException("Condition minLOD <= maxLOD does not meet.");
-            MinLOD = minLOD;
-            MaxLOD = maxLOD;
-            AvailableMaxLOD = availableMaxLOD;
         }
     }
 }
