+++ Runtime/CityImport/Config/PackageLoadConfigs/PackageLoadConfigExtendable.cs
@@ -1,4 +1,5 @@
 using System;
+using PLATEAU.CityImport.Config.PackageLoadConfigs;
 using PLATEAU.PolygonMesh;
 
 namespace PLATEAU.CityImport.Config
